requirements.txt
streamlit
tensorflow
numpy
pillow
json5
keras
matplotlib
scikit-learn
opencv-python
gunicorn
